Update python_ex228.py
pg 121 in the self-taught programmer. Should import "module1" (instead of "hello") according to instructions and for reproducing the autorun error with imported modules. Importing "hello" does not produce the same error mentioned in the prior example (see previous pull request), therefore to resolve the error, will also need to call "module1" in this example.